% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinf_bayes.R
\name{zinf_bayes.stanreg}
\alias{zinf_bayes.stanreg}
\alias{zinf_bayes}
\title{Create a Bayesian Zero-Inflation Model}
\usage{
\method{zinf_bayes}{stanreg}(mody, modp, ...)
}
\arguments{
\item{mody}{A rstanarm::stan_lmer() object.}

\item{modp}{A rstanarm::stan_glmer() object.}
}
\value{
An object of class `zinf_bayes`.
}
\description{
`zinf_bayes()` is a generic function used to wrap together the two individual
models in a zero-inflation model. The function invokes particular methods
such as `summary` and `predict` that allow it to be used within typical
R modeling workflows
}
\details{
Importantly, this generic does no model building itself. The
pre-existing rstanarm modeling functions are already so robust that it is
sensible to lean on the quality of that software instead of trying to
reinvent it within our own function
}
