% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinf_freq.R
\name{zinf_freq.merMod}
\alias{zinf_freq.merMod}
\alias{zinf_freq}
\title{Create a Frequentist Zero-Inflation Model}
\usage{
\method{zinf_freq}{merMod}(mody, modp, ...)
}
\arguments{
\item{mody}{A lme4::lmer() object.}

\item{modp}{A lme4::glmer() object.}
}
\value{
An object of class `zinf_freq`.
}
\description{
zinf_freq is a generic function used to wrap together the two individual
models in a zero-inflation model. The function invokes particular methods
such as `summary` and `predict` that allow it to be used within typical
R modeling workflows
}
\details{
Importantly, this generic does no model building itself. The
pre-existing lme4 modeling functions are already so robust that it is
sensible to lean on the quality of that software instead of trying to
reinvent it within our own function
}
